api/PromptFlow/ChatGpt/flows/standard/requirements.txt
promptflow_vectordb[azure]
azure-cosmos==4.4.0b1